third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/style-sharing-host-attribute-expected.txt

CONSOLE ERROR: line 18: Uncaught TypeError: roots[i].createShadowRoot is not a function
Don't share style between two hosts with different attribute matching for :host([])

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

This text should be green
